refactors venues page into separate components and hook
